created helper for get menus and status
