+++ App/Controllers/Front/IndexController.php
@@ -2,7 +2,6 @@
 
 namespace App\Controllers\Front;
 
-use App\Repositories\Front\IndexRepository;
 use Core\View;
 use Core\Controller;
 
@@ -15,29 +14,11 @@ class IndexController extends Controller
 {
 
     /**
-     * @var
-     */
-    protected $menu_repo;
-    /**
-     * @var IndexRepository
-     */
-    protected $repo;
-
-    /**
-     *
-     */
-    public function __construct()
-    {
-        $this->repo = new IndexRepository();
-    }
-
-    /**
      * @throws \Exception
      */
     public function indexAction()
     {
-        $menus = $this->repo->getMenus();
-        View::render('Front/index.php', ['menus' => $menus]);
+        View::render('Front/index.php');
     }
 
 }
