+++ App/Repositories/Admin/MenuRepository.php
@@ -4,6 +4,7 @@ namespace App\Repositories\Admin;
 
 use App\Models\Menu;
 
+
 /**
  * Class MenuRepository
  * @package App\Repositories\Admin
@@ -22,6 +23,8 @@ class MenuRepository
     public function __construct()
     {
         $this->model = new Menu();
+        $this->page_repo = new PageRepository();
+        $this->post_repo = new PostRepository();
     }
 
     /**
@@ -81,6 +84,27 @@ class MenuRepository
     public function updateMenuSortOrder($sort_order)
     {
         return $this->model->updateMenuSortOrder($sort_order);
+    }
+
+    public function getMenus()
+    {
+        $allMenu =$this->getAll();
+        $menuData = [];
+        foreach ($allMenu as $value) {
+
+            $menuData[$value['id']]['name'] = $value['name'];
+            if ($value['type'] == 1) {
+                $data = $this->post_repo->getPostById($value['link']);
+                $menuData[$value['id']]['slug'] = $data['0']['slug'];
+            } elseif ($value['type'] == 2) {
+                $data = $this->page_repo->getPageById($value['link']);
+                $menuData[$value['id']]['slug'] = $data['0']['slug'];
+            } elseif ($value['type'] == 3) {
+                $menuData[$value['id']]['slug'] = $value['link'];
+            }
+        }
+        return $menuData;
+
     }
 
 }
