+++ public/theme/admin/assets/css/custom.css
@@ -50,6 +50,7 @@
     background-color: #fff;
     border: 1px solid #ddd;
     color: #226b92;
+    cursor: move;
 }
 .list-group-item.active .list-group-item-heading, .list-group-item.active .list-group-item-heading>.small, .list-group-item.active .list-group-item-heading>small, .list-group-item.active:focus .list-group-item-heading, .list-group-item.active:focus .list-group-item-heading>.small, .list-group-item.active:focus .list-group-item-heading>small, .list-group-item.active:hover .list-group-item-heading, .list-group-item.active:hover .list-group-item-heading>.small, .list-group-item.active:hover .list-group-item-heading>small {
     color: inherit;
